add --random parameter to match
rationale: when testing bots against each other which do **not** emit
random actions but always derive their actions from the map state,
it can lead to somewhat unfair comparisons. if the simulation
injects some randomness it might help to better compare the algorithms
without needing to add the randomness to the bots themselves.